Add spinners
