--- src/components/DeploymentView/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "DeploymentView",
-  "version": "0.0.0",
-  "main": "./DeploymentView.js"
-}
